source/requirements.txt
# Function dependencies, for example:
# package>=version
python-dateutil
httplib2
pytz